refactor(app): centralize extension setup
